Create reviewSummary component and copy over data
